Fix `SDKS=somesdk make build` (#19996)
Previously `make build` depended on `dist`, which then recurses into the
subprojects. This meant that `SDKS=somesdk make build` would run the
`build` target for the CLI and the `dist` target for the SDK.

With this PR, `SDKS=somesdk make build` now runs the `build` target for
both the CLI and SDK.

This is particularly useful for the Node.js SDK, where `build`
transpires the TypeScript files. Previously `SDKS=nodejs make build`
would mysteriously not pick up changes to the TypeScript files.