Fix the AI prompt flow for pulumi new (#19998)
Fix https://github.com/pulumi/pulumi/issues/19994 which is a regression
from https://github.com/pulumi/pulumi/pull/18627

The change is essentially moving the block which assigns
`args.templateNameOrURL` above the call to `cmdTemplates.New` which uses
`args.templateNameOrURL`. With the existing order,
`args.templateNameOrURL` set to the AI URL was ignored.

The rest of the change is refactoring the AI prompting to a separate
function and a test that mocks that function to a hard-coded value.