Implement ResourceHooks in the engine (#19874)
Implement lifecycle hooks in the engine. This does not yet expose hooks
in the SDKs.

There are currently 6 types of resource hooks:

- `BeforeCreate` 
- `AfterCreate`
- `BeforeUpdate`
- `AfterUpdate`
- `BeforeDelete`
- `AfterDelete`

Hooks can throw an error. For hooks that execute before a step, an error
causes the step to fail. For hooks that execute after a step, we log a
warning.

By default hooks don't run on dry runs, but can be configured to do so.

To bind hooks to a resource, the `hooks` `ResourceOption` takes
`ResourceHookBinding`.

Under the hood, hooks are `Callback`s that are registered with the
engine using the `RegisterResourceHook` RPC. The engine keeps track of
all the known hooks in the `ResourceHooks` register and will call the
language runtime's callback server to invoke the hooks at the
appropriate times.