Fix ListStacks to use the same default-org logic as loading a single stack (#20010)
Fixes https://github.com/pulumi/pulumi/issues/19598

There was an inconsistency between GetStack and ListStacks on deciding
what the default org was. This could result in stacks from GetStack
thinking the default org was A, and stacks from ListStacks thinking the
default org was B.

When rendering out the names this could result in the same stack being
rendered different. GetStack for "A/proj/stk" would print "stk" because
it thought A was the default org, while ListStack would render
"A/proj/stk" because it thought B was the default org.

When used in stack select code this could result in an error from the
survey library as we gave it a list of stacks (including "A/proj/stk")
but told it the default was just "stk" (which wasn't in the list).