Fix flaky TestDBRParallel (#20044)
Fixes https://github.com/pulumi/pulumi/issues/20042

Rather than relying on bool flags being set in the right order we now
actually wait on `WaitGroups`. This does mean if we break this test it
will probably manifest as a timeout, but currently these wait groups are
guarnteed to be triggered in the right order, because that's what
they're testing for.