Automatic ESM configuration (#19980)
We support native ESM mode (meaning modules are not transpiled to
CommonJS, among other things), but it requires some explicit
configuration, see
https://www.pulumi.com/docs/iac/languages-sdks/javascript/#native-esm-support

With this PR, we attempt to configure everything automatically.

If we detect manual configuration, we fallback to the old code path, to
avoid messing with any custom configuration.

---------

Co-authored-by: Fraser Waters <fraser@pulumi.com>