Fix `ResourceV3.RefreshBeforeUpdate` yaml struct tag property name (#20045)
This changes fixes the wrong property name in the yaml struct tag for
`ResourceV3.RefreshBeforeUpdate` (copy/paste issue), which I noticed
looking at the code. I don't think it impacts anything in practice
(yaml).

Note: I looked into enabling a linter to catch things like this
([Tagliatelle](https://github.com/ldez/tagliatelle) via golangci-lint),
but we have a bunch of existing casing/naming inconsistencies (50+) that
would have to be ignored, so punted on it for now.