Support reading v4 checkpoints/deployments (#20046)
This change adds support for reading deployments with a schema version
of 4.

Additionally, this change adds support for a new `Features` field that
can be specified on a deployment/checkpoint to indicate required
features. If a CLI does not support one of the specified features, it
will produce an error when reading the deployment/checkpoint. This way,
we can add support for new required features in the future without
having to bump the schema version.

Part of #19705

There will be two subsequent changes after this:
1. Reading/round-tripping v4 support in Pulumi Cloud
2. Writing v4 from CLI