replace assert.NotNil with require.NotNil
Similar to https://github.com/pulumi/pulumi/pull/19988, replace
assert.NotNil with require.NotNil.  When we check that something is
not nil, we very frequently the use the same object in the next step.
If it is nil, and we only use `assert.NotNil`, we will just record the
error and then fail in the subsequent step.

Instead with `require.NotNil`, we fail immediately, recording the
error, but not crashing the test suite at the same time.