Enable lifecycle tests on Windows
Lifecycle tests on Windows are currently disabled in CI. After some
discussion and investigation, it appears that this might be due to
failures caused by line ending differences in Windows-generated outputs
and POSIX-generated snapshots. If this is the case, we might be able to
run the tests simply by disabling comparing snapshots, which would be a
great start. This change does just that, by setting the
`SkipDisplayTests` flag for a test to true if the OS is Windows,
regardless of the setting passed in.

Part of #19675