stop lifecycletests form deadlocking when a `require` fails in the program function (#20062)
Since
https://github.com/pulumi/pulumi/pull/19837/files#diff-984b3279d61b588030474831898bb76d093d863ca2d0015bb081d842602a9ab7R841,
we're waiting for the program to finish running before shutting the
engine down.

This is useful in regular operation, however it introduces a small issue
in the lifecycletests. Because programs can be canceled using by the
`require.*` family of functions, we sometimes don't "finish" running the
program correctly, if there are test failures. In those cases, we
currently just hang, which makes the tests much harder to debug. Fix
this.

The inline comment describes the issue a bit more carefully.