Simplify determining the Node.js entrypoint and respect package.exports (#19985)
This was modeled on the internal Node.js code, which has since changed
quite a bit. We can simplify this to match our supported versions of
Node.js.

See
https://github.com/nodejs/node/blob/v20.x/lib/internal/modules/run_main.js#L23

We know we're in ESM mode, so this boils down to
`url.pathToFileURL(path.realpath(path.resolve))`.

We now also respect `package.exports`.