Fix `package add` command when using uv 0.8.0 or later
Starting with version 0.8.0, uv will automatically add packages in subdirectories as workspace members. However the generated SDK might not have a `pyproject.toml`, which is required for uv workspace members. To add the generated SDK as a normal dependency, we can run `uv add --no-workspace`, but this flag is only available on version 0.8.0 and up.g