Fix before exit handler (#20070)
The aws provider has a handler that creates `BucketNotification` objects
in a `beforeExit` handler, see
https://github.com/pulumi/pulumi-aws/blob/df45f46766be1d304a5fcf7d6dc192846f7433a8/sdk/nodejs/s3/s3Mixins.ts#L187

`beforeExit` handlers are run in FIFO order, and since we register
[ours](https://github.com/pulumi/pulumi/blob/dbfb8f2ccd8b11fb66ccbe5cced5abb324a554f5/sdk/nodejs/cmd/run/index.ts#L84)
before the program is loaded, it runs before the AWS provided one. This
is problematic, because we call `SignalAndWaitForShutdown` in the
handler, but then a resource registration comes in afterwards, which
will fail.

To fix this, we employ a bit of a hack to ensure our handler runs last,
and in fact only runs after all the async work spawned by the first set
of handlers has completed.

A user or library could still attempt the same trick with `nextTick`,
but we are mostly powerless to fix this. Even if we inspect the list of
handlers after the program returns, there could be pending async work
that will later register more handlers, etc. Generally it seems unlikely
that someone would want to register resources in `beforeExit`, and of
course we did this to ourselves in the AWS provider.

Fixes https://github.com/pulumi/pulumi/issues/20068