replace assert.Len with require.Len (#20076)
Similar to https://github.com/pulumi/pulumi/pull/19988 and
https://github.com/pulumi/pulumi/pull/20055 previously, replace
`assert.Len` with `require.Len`. The reasoning is similar to those other
two PRs, namely that after `assert.Len` we often check the elements of
arrays. If the length isn't correct, we'll panic because of trying to
access an array index out of range.

Note that this isn't always the case e.g. when asserting the length is
0, but overall I think it's better to just use `require.Len` everywhere,
so we don't need to think about it in every case.

This should be the last PR in this series.