Don’t use automatic ESM mode when `—import` or `—require` is set
Similar to how we skip automatic ESM mode when there is a `--loader` flag, skip it in the presence of `--import` or `--require` flags. For example [tsx](https://tsx.is) is often used with `--import tsx`.

Usually we wouldn’t hit this case, because we require a recent version of `ts-node` to activate automatic ESM mode, and when using `tsx` there’s no reason to have `ts-node` explicitly installed. It might still be present though, either directly in `package.json` for some other use, or as an indirect dependency.

This could lead to a setup where code would first run though `tsx`, stripping the types, and then through `ts-node/esm`, which could then potentially fail with a type error since it was now dealing with a lot of `any` and `unknown` types.