Don't print `Output.__str__` errors when using `_safe_str` (#20094)
In #18903 we introduced the `_safe_str` function, which avoids throwing
errors when stringifying `Output`s in the core Python SDK. When the
`PULUMI_ERROR_OUTPUT_STRING` environment variable is set, `_safe_str`
produces a fallback string (`Output[T]`) instead of throwing an error.
`_safe_str` is then used in log statements in the core SDK so that we
can produce debug information without blowing up user programs depending
on the settings they have enabled.

Ironically, this means that if `PULUMI_ERROR_OUTPUT_STRING` is *not*
set, the logs we produce are worse! Rather than containing nice neat
`Output[T]` strings, they contain the classic `Calling __str__ on an
Output[T] is not supported...` spiel, which makes for a lot of noise.
What we really want is for `_safe_str` to *always* produce the string
`Output[T]`, since the places we use it have no use for the warning.
This change does just that, leaving in the "catch the error" code path
in case we end up using `_safe_str` on objects that contain outputs
later on.