Fix callback server hang
The running callbacks server keeps the Node.js eventloop
active, so we will never exit the process. Normally we would
shutdown the server when we're done, however we don't have a
good way to tell when that is the case. Consider a user
provider Pulumi program:

```javascript
setTimeout(() => new random.RandomPet(
     "username", {}, { transforms: [...] }), 2000);
```

When we import the user program (module really), it will
return immediately, but we have to wait for the timeout to
trigger and then the resource registration to complete.

Usually the best indication that no more work is outstanding
is `process.on("beforeExit", ...)`, which will be emitted
when there is no more work waiting in the eventloop. However
if we have a server running, we'll never get to that event.

To break out of this catch-22, we `unref` the server, and the
sessions created by the server.