unflaky analyzer canellation test
Increase the sleep time here, similar to what we do in language
plugins.  See also [the conversation in that PR](https://github.com/pulumi/pulumi/pull/20007#discussion_r2206677998).

Also while we're there, assert on the content of the error.

Note that this test can still flake when a computer is very busy
running stuff, but at least this seems better than before, and less
flaky for me locally.