Fix data race in events system (#20105)
Fixes a data race between the event system trying to deepcopy step
metadata objects (which contain `*resource.State` objects) and the step
executor that could be trying to set fields on those state objects.

The step executor was correctly trying to lock/unlock the state objects,
but the event system wasn't.

This was a fairly benign race, I don't think we would of ever really
seen user issues because of it but it would trigger the go race detector
and fail tests.