Fix callback server hang & delete hooks in Node.js (#20109)
The running callbacks server keeps the Node.js eventloop
active, so we will never exit the process. Normally we would
shutdown the server when we're done, however we don't have a
good way to tell when that is the case. Consider a user
provided Pulumi program:

```javascript
setTimeout(() => new random.RandomPet(
     "username", {}, { transforms: [...] }), 2000);
```

When we import the user program (module really), it will
return immediately, but we have to wait for the timeout to
trigger and then the resource registration to complete.

Usually the best indication that no more work is outstanding
is `process.on("beforeExit", ...)`, which will be emitted
when there is no more work waiting in the eventloop. However
if we have a server running, we'll never get to that event.

To break out of this catch-22, we `unref` the server, and the
sessions created by the server.

Other than transforms, this is also an issue for delete hooks.
For delete hooks it also fixes the issue that we shutdown
the callback server early, causing the hook to fail with
a GRPC error (because the callback server was closed and then
later started up again on a different port).

Fixes https://github.com/pulumi/pulumi/issues/20099