Don’t report leaked promises when we exit abnormally
After we started using `SignalAndWaitForShutdown` in https://github.com/pulumi/pulumi/pull/19946, we now have more situations were the Node.js program shuts down cleanly, even though the deployment is failing.

This can cause the promise leak debugging message to be shown, but in the presence of other errors this is just noise. Track early monitor termination in the state, and use that information to decide whether we should show the debug message or not.