Don’t report leaked promises when we exit abnormally (#20128)
After we started using `SignalAndWaitForShutdown` in
https://github.com/pulumi/pulumi/pull/19946, we now have more situations
were the Node.js program shuts down cleanly, even though the deployment
is failing. This can cause the promise leak debugging message to be
shown, but in the presence of other errors this is just noise, since
Pulumi will have reported the original error to the user.

Track early monitor termination in the state, and use that information
to decide whether we should show the debug message or not.

Note: it is rare now for a program to exit on its own with a status code
0 and thus potentially trigger the promise leak message. The user would
have to call `exit(0)`. Prior to
https://github.com/pulumi/pulumi/pull/5347 the SDK itself would call
`exit(0)` when the monitor was shutdown.

---------

Co-authored-by: Fraser Waters <fraser@pulumi.com>
Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>