Fix delete hooks in inline automation API programs
The inline automation API language runtime needs to call `SignalAndWaitForShutdown` when the program complets.

Note that for inline programs the lifetime of the program is tied to the program function execution. When the program function returns, we wait for any queued work and then shutdown. Work that gets queued later will not be executed (or might if it manages to race the shutdown), see https://github.com/pulumi/pulumi/issues/20114