Fix hang when waiting for program that would never complete
The engine knows when a resource registration fails and cancels the deployment. An SDK might export such a failed resource as a stack output. In Node.js and Python, a failed resource registration causes the properties of the resource to never resolve. If the SDK code is awaiting on the resource to resolve, we might hang forever: the engine waits for the program to complete, and the program waits for the resource to resolve.

The fix is to not wait for the program completion (via SourceIterator.Cancel) if we’re already in the process of exiting due to another error.

Fixes https://github.com/pulumi/pulumi/issues/20151