Generate same steps in delete code (#20061)
Noticed this while looking into other delete bugs. When resources are
removed from a program, but not targeted they wouldn't get a goal state,
but would also skip the delete step. They _should_ have shown up as
sames, but because we never generated actual steps for them the display
layer just didn't know about them.

This also resulted in the fixes we had to put in recently to add skipped
delete resources to the `sg.deployment.news` map. That always looked a
bit suspect and is now no longer needed, these get added to `news` via
the same steps just like any normal resource.

This did require one suspect change in the step executor where we can no
longer use the presence of a goal state to indicate a resource is
managed (because these skipped deletes don't have goals) but flipping
that if to just check for a ReadOp first seems fine as only external
resources will use that operation.