Fix hang when waiting for program that would never complete (#20153)
The engine knows when a resource registration fails and cancels the
deployment. A program might export such a failed resource as a stack
output. In Node.js a failed resource registration causes the properties
of the resource to never resolve. If the SDK code is awaiting on such a
stack output to resolve, as is the case for the Automation API, we might
hang forever: the engine waits for the program to complete, and the
program waits for the resource to resolve.

The fix is to not wait for the program completion (via
`SourceIterator.Cancel`) if we’re already in the process of exiting due
to another error.

Fixes https://github.com/pulumi/pulumi/issues/20151

---------

Co-authored-by: Fraser Waters <fraser@pulumi.com>