Fix panic in up --refresh --run-program in dependency graphs (#20147)
When running `up` with `--refresh --run-program` we end up with three
state objects per resource (as opposed to the usual case of two). We
have the original state we loaded from the snapshot (A), then we have
the refreshed state after calling Read on the provider (B), and then we
have the new state from the program goal and the same/update step (C).

The dependency graph logic expected to be able to do pointer comparisons
to resolve state lookups but if we create the dependency graph from the
'A' states, later when we try to look up the dependencies of B, which is
now the 'old' state from a step, it doesn't match by address.

To fix this when we create one of these middle states (B) for a refresh
operation we add an alias to the dependency graph for it, so later when
we get the steps 'old' resource state pointer it just aliases back to
the original old state in the dependency graph.