Add get_root_directory to root pulumi module (#20172)
This got added back in https://github.com/pulumi/pulumi/pull/18595 but
you had to use it via `pulumi.runtime.get_root_directory()`. There isn't
really anything else that we expect users to look at in the runtime
module, so this copies it to the root pulumi module so it's now just
`pulumi.get_root_directory()`.