Fix flaky TestNodejsPromiseMessage (again) (#20171)
I attempted to fix this in https://github.com/pulumi/pulumi/pull/20144,
but misunderstood how promise leak detection works. We need a pending
`debuggablePromise` when the process exits to trigger the message.

Something like `sleep(60000).then(() => { new Named("res”); })` will not
create a `debuggablePromise` until we hit the resource constructor.

The reason the test passed previously was that it would usually, but not
always, hit `process.exit` while debug logs (which uses
`debuggablePromise`, since it's an RPC) were being sent.