Add getRootDirectory to root pulumi module (#20173)
Same fix as https://github.com/pulumi/pulumi/pull/20172 but for nodejs.

This got added back in https://github.com/pulumi/pulumi/pull/18595 but
you had to use it via `pulumi.runtime.getRootDirectory()`. There isn't
really anything else that we expect users to look at in the runtime
module, so this copies it to the root pulumi module so it's now just
`pulumi.getRootDirectory()`.