use the testify framework more in lifecycletests (#20175)
I ran into this while working on
https://github.com/pulumi/pulumi/pull/20174. Testify returns the whole
stack trace rather than just showing the log, which is often useful to
see what actually failed. Use that framework here consistently, so we
get better test output.

There's also a bug here, where we don't fail if we encounter an event we
don't expect, so fix that as well.