Enable `pulumi import` to always `--generate-resources`
When using `pulumi import` to import resources `--from` a converted
state file, Pulumi will write out a JSON-encoded set of resources if
things go wrong so that the user can then try to fix things up and try
again using the `--file` option instead. There are cases however where
the JSON-encoded set of resources is useful even when an import
succeeds. This change adds the `--generate-resources <path>` option to
`pulumi import` so that users can force the generation of resources JSON
regardless of whether the import succeeds or fails. It also fixes a
small double-`%s` error in one of our diagnostic messages, since that's
the kind of bonus content we live for.