record test output for previews in lifecycletests (#20174)
In the lifecycletests, we check the test output against "golden"
files.  However we return early if this is a dry run/preview.  We
still have output in these cases, so we should also check against the
golden files there.  Make it so.

As I looked into @Frassle's comment in
https://github.com/pulumi/pulumi/pull/20161#pullrequestreview-3070699684,
I realized that we don't save the display data for previews at all. I
don't recall a reason for not doing so, presumably this was mostly by
accident. Fix this and also record the test output for previews in
lifecycletests, and compare against it later.

The first commit here is all the code changes, while the second one
actually adds the test data.