preview: show outputs if they changed (#20161)
In preview, we're currently not showing outputs that changed, unless
they have a detailed diff attached to them. So for example if someone
exports "bucket.id", but later changes that to a string, it currently
doesn't show up in the diff preview. This is rather confusing, and we
should be showing these outputs.

Fixes https://github.com/pulumi/pulumi/issues/20087