Fix cancellation when running via some CLI wrapper tools (#20195)
In https://github.com/pulumi/pulumi/pull/19603 we changed Ctrl-C
handling to signal to Pulumi’s process group instead of just the Pulumi
CLI. This mostly work, however when running via the 1Password CLI or our
own ESC, the Pulumi CLI is not the process group leader, and no process
group matching the Pulumi CLI PID exists. This lead to the SIGINT being
sent nowhere (with the error ignored).

We do not actually need to signal our process group, as Pulumi already
tracks any subprocesses it starts, and shuts them down properly. The fix
is then to signal just the Pulumi CLI.

A different approach would be to always create a process group for the
Pulumi CLI, this was implemented in
https://github.com/pulumi/pulumi/pull/20189 but we chose the slightly
more versatile approach here.

Fixes https://github.com/pulumi/pulumi/issues/20154

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>