[chore] Add bun as a .mise.toml dependency.
While running `make` on the project, some of the tests on
`sdk/nodejs/npm/manager_test.go` started failing as `bun` was not
installed on my system.

I considered introducing go test tags to gate whether these tests should
run or not based on `bun` being present. However, I could not find
whether we are using that mechanism for anywhere else. Hence, I decided
to simply add the dependency to the latest version of `bun` to
`.mise.toml`.