Add `--remove-backups` flag to `pulumi stack rm` (#20203)
This change adds a new `--remove-backups` flag to the `pulumi stack rm`
command. This flag will remove any backups of the stack if using the DIY
backend. It is a no-op if using the cloud backend.

Fixes #9474

Question: Assuming we're happy with exposing this via this new flag, do
we also want to expose it from Automation API (my default assumption is
"yes")?