Move ID="" logic from grpc layer to refresh step (#20202)
Prompted by https://github.com/pulumi/pulumi/issues/20186

The logic that `ID=""` should be treated as a delete was squeezed into
the grpc layer by setting outputs to nil if ID was blank.

This PR moves that logic up to the refresh step, so it's consistent
across all lifecycle tests. A number of tests and the builtin provider
needed updating for this.