Fix TestRunningViaCLIWrapper
Use github.com/creack/pty, not github.com/kr/pty. The latter is a thin wrapper on top of the former and deprecated.

The for loop was setup up wrong, it was meant to be running for up to 3 minuites, but only ran for 300 ms. In CI, which is typically running slower than locally, we’d send `SIGINT` before we’d even started pulumi.