Restore terminal mode on panic (#20224)
Pulumi sets the terminal to raw mode, and on exit restores the terminal
to its previous state. However when we crash due to a panic, we don’t
restore the state.

By deferring the `Close` call, we’ll call the terminal’s `Close` method
when we panic.

This only works if the panic comes from the same goroutine as the
display code. If we panic in another goroutine, we won’t call the
deferred method, and we still won't restore the state.