Test and fix a bug in `up --refresh --run-program` with replaced resources (#20230)
Fixes https://github.com/pulumi/pulumi/issues/20217

This fixes a snapshot integrity error when running `up --refresh
--run-program` that replaced a resource. The problem was that `--refresh
--run-program` causes three `*resource.State` instances to be created
during the deployment run. First we have the original state from the
snapshot, then the state created by the refresh step, then the final new
state from the goal state.

When running deletes at the end of the deployment we need to match the
original states to the refreshed states, as those intermediate states
are the ones that will have had `Delete` set on them and will currently
be in the snapshot (the original being replaced by the refresh state by
the `RefreshStep`).

---------

Co-authored-by: Julien <julien@caffeine.lu>