Calculate PluginInfo.Size on demand
This can be somewhat slow for when there’s for example large node_modules folders in the plugins. In practice we never care about the size, except when displaying it in `pulumi plugin ls`.

This is a breaking API change. I checked https://github.com/pulumi/pulumi-terraform-bridge and it compiles fine with this change, it never uses PluginInfo.Size.