Fix race when reading plugin stdout
When running a binary plugin, we used `cmd.Stdout/StderrPipe` to get the output from the binary. These pipes use copying goroutines to move the copy the data around, and when the process exits, the pipes are immediately closed. You must read from the pipe before `cmd.Wait()` is called, otherwise you will miss data.

From the documentation:

> It is thus incorrect to call Wait before all reads from the pipe have completed

However, ExecPlugin returns an already started plugin and runs a goroutine that waits for the plugin to exit:

```go
	go func() {
		err := cmd.Wait()
		if err != nil {
			wait.Reject(err)
		} else {
			wait.Fulfill(struct{}{})
		}
	}()
```

If a plugin exits very quickly, this can lead to us trying to read from an already closed pipe.

This happens in the test `TestNonZeroExitcode`, where we see an error:

```
        	Error:      	Error "could not read plugin [/tmp/TestNonZeroExitcode36998759/001/test-plugin-exit]: read |0: file already closed" does not contain "test-plugin-exit]: exit status 1"
```

We can fix this by using `io.Pipe`. These pipes are synchronous and writes from the plugin while block until we set up a reader.