Fix component providers when using pnpm
During schema analysis, we need to determine the package name for resource references. To do this, we look at the import path for a given module, and attempt to extract the name from there. The regex assumed the simple case of `node_modules/@org/name` or `node_modules/name`, but node_modules can be nested. The right most package name is the one we’re looking for.

Once that was fixed, we ran into a 2nd issue. pnpm will not run postinstall scripts by default, so `pulumi package add …` would not compile the generated TypeScript to JavaScript.

Fixes https://github.com/pulumi/pulumi/issues/20244