Add test for paramterizing a provider twice (#20243)
This will be useful for confirming that
https://github.com/pulumi/pulumi-dotnet/issues/664 has the right fix.

As part of this I had to change the `Providers` lists in the conformance
tests to be thunks (i.e. `[]func() plugin.Provider` instead of just
`[]plugin.Provider`) so that we can instantiate two copies of the
parameterised provider (as it's stateful). We had a TODO to do this
anyway for memory consumption, so double win.