Remove Pwd from diag.Sink and delete TestSink (#20268)
The TestSink wasn't used anywhere, I'll see about adding a new one
suitable for tests to the diag package directly because I'm sure we have
a few around different packages. Having it in testutil was awkward
because testutil depends on "backend" so it was easy to get package
loops.

The 'Pwd' field on the format options wasn't used anywhere, so also
deleting that.