Remove Stringify from Sink interface (#20273)
While the Stringify function is used, and both real implementations of
Sink have a Stringify function, there wasn't anything that made use of
the fact that it was on the interface.

Remove it from the interface so we now have a simpler interface that
will be easier to write a mock type around. As opposed to the old
TestSink which had to embed the defaultSink in it just to get access to
Stringify.