Use nullish coalesce operator for required property checks and state and args accessing whenever sensible. (#20266)
The generated TypeScript code currently uses checks like `x ? x.y :
undefined` in places which are massively repeated in generated code.
By using nullish coalesce operators `?.` and `??` more aggressively, the
generated code is more compact, and thus, nicer to read.

All of the actual changes are to one file: `pkg/codegen/nodejs/gen.go`.
The remaining changes are test data updates, or changelog entries.
This PR should not have any functional effects - it only changes the
syntax of generated TS files to be more compact.

This PR is related to #20261 , implementing one part of it.
The original PR will need test case added to it, but these particular
changes do not require that, so this is why I created separate PR for
them.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>