Fix a bug with `destroy --exclude-protected --remove` and empty stacks (#20283)
When destroying an empty stack with `--exclude-protected --remove` you
would expect it to do a no-op destroy and then remove the stack, instead
you would get an error:

```
There were no unprotected resources to destroy. There are still 0
protected resources associated with this stack
```

This fixes the command to just carry on for empty stacks and do the
destroy and stack removal.