add an ErrorEvent engine event
Add an ErrorEvent type that the engine can use to send error events to
the service. This event type is meant for internal errors that can
help debug issues, but is not meant to be user visible. In particular
it's going to be used for Journaling, and comparing the journal and
the manager snapshot and sending an engine event if they are different.