add an ErrorEvent engine event (#20291)
Add an ErrorEvent type that the engine can use to send error events to
the service. This event type is meant for internal errors that can help
debug issues, but is not meant to be user visible. In particular it's
going to be used for Journaling, and comparing the journal and the
manager snapshot and sending an engine event if they are different.

I *think* this is the first step to introduce a new engine event that
the service wants to get, as `pulumi/pulumi-service` then pulls in the
apitype from `sdk/go/common`.