Don't error for missing old paths in ignoreChanges (#20278)
Fixes https://github.com/pulumi/pulumi/issues/20227

Historically when using ignoreChanges if you tried to ignore a path that
didn't have old data we would error. For example if your old input state
looked like `{foo: "hi"}` and you tried to ignore the path "bar" we
would error `"cannot ignore changes to the following properties because
one or more elements of the path are missing: bar"`.

You got this error regardless of if the new input state had the path or
not. While cases like the above are unusual it's really easy with
wildcard paths and arrays to hit this, because when ever you added an
item to the array that index would not be in the old state and you'd get
an error about it (see #20227).

However I don't think this actually needs to be an error case at all.
`ignoreChanges` is simply instructions to use the old inputs rather than
the new inputs, but if there aren't any old inputs we can just use the
new input data as is. This is what we do on first create for example.
We'll still print an info diagnostic to let the user know that the
ignoreChange path didn't apply anything, should help with cases of
people making typos here, but I don't think it ever needs to be a full
error.

If we had full schemas loaded in the engine (yup that old chestnut
again), we probably could error for paths that aren't in the schema.
Which would be a good way of catching typos, without breaking cases like
the changing array size above.