[import] Use mapped name instead of logical name for resource references during import code generation (#20300)
Fixes #20257

During import code generation, when we emit resource blocks for
_dependant_ resource references, we always opted to use the logical name
of that resource to create the reference variable i.e.
`<logical-name>.id` instead of `<name>.id`. This leads to PCL parsing
errors when the logical name is not a valid identifier, for example `My
Bucket & Content.id` (fails to parse).

This PR fixes the issue by correctly using the names table to derive the
lexical name used when creating the variable reference