[proto] Add stack trace to specific requests
The following requests may now contain stack traces from the caller:

- ReadResourceRequest
- RegisterResourceRequest
- ResourceInvokeRequest
- ResourceCallRequest

All of the above are sent by resource monitor clients. Their stack
traces may reference code in the user's Pulumi program or code in a
multi-language component in a provider. In addition to the caller's
stack trace, all of the above besides ResourceInvokeRequest may also
contain a handle for a parent stack trace. That handle is used by the
engine to concatenate stacks for calls that cross plugin boundaries.

The following requests may now contain a handle to a stack trace:

- CallRequest
- ConstructRequest

These requests are sent to providers by the engine. The stack trace
handle should be recorded by the provider so that subsequent requests to
the resource monitor can contain the handle for stack concatenation.