[apitype] Capture stack traces in state
Just what it says on the tin. Stack traces are captured using two new
types: one for the stack trace and one for each frame it contains. This
will allow us to add additional information to traces and frames in the
future if we so choose.

Note that the "sourcePosition" property of a stack frame is optional.
This allows us to support stacks that contain frames that were missing
position information at the time they were captured.