[engine] Track and record stack traces
These changes add the necessary machinery to the engine to capture and
persist stack traces. This includes support for concatenating stack
traces that cross plugin boundaries.

Stack concatenation requires that the resource monitor is able to track
call stacks across plugins. Call stacks may cross plugins as part of a
call to Call or the registration of a remote component. In both cases,
the monitor extracts the incoming stack from the request and associates
it with a handle (currently a UUID, but could be anything sufficiently
unique). This handle is then passed on to the associated Call or
Construct call made by the monitor in response to the original request.
When the monitor receives a Call/Read/Register that contains a parent
stack handle, it extracts the handle's associated stack trace and
prepends it to the stack trace from the request.